backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aniso8601==9.0.1
    # via flask-restful
boto3==1.26.115
    # via -r requirements.in
botocore==1.29.115
    # via
    #   boto3
    #   s3transfer
click==8.1.3
    # via flask
flask==2.2.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-mysql
    #   flask-restful
flask-cors==3.0.10
    # via -r requirements.in
flask-jwt-extended==4.4.4
    # via -r requirements.in
flask-mysql==1.5.2
    # via -r requirements.in
flask-restful==0.3.9
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.7.0
    # via -r requirements.in
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector==2.2.9
    # via -r requirements.in
pyjwt==2.6.0
    # via flask-jwt-extended
pymysql==1.0.3
    # via flask-mysql
python-dateutil==2.8.2
    # via botocore
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via flask-restful
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   flask-cors
    #   flask-restful
    #   python-dateutil
urllib3==1.26.15
    # via botocore
werkzeug==2.2.3
    # via
    #   flask
    #   flask-jwt-extended
